moved influxdb write result job
