+++ app/Jobs/Ookla/ProcessSpeedtestBatch.php
@@ -3,7 +3,6 @@
 namespace App\Jobs\Ookla;
 
 use App\Jobs\CheckForInternetConnectionJob;
-use App\Jobs\InfluxDBv2\WriteSpeedtestResult;
 use App\Models\Result;
 use Illuminate\Bus\Batch;
 use Illuminate\Contracts\Queue\ShouldQueue;
@@ -36,7 +35,6 @@ class ProcessSpeedtestBatch implements ShouldQueue
                 new RunSpeedtestJob($this->result),
                 new BenchmarkSpeedtestJob($this->result),
                 new CompleteSpeedtestJob($this->result),
-                new WriteSpeedtestResult($this->result),
             ],
         ])->catch(function (Batch $batch, ?Throwable $e) {
             Log::error(sprintf('Speedtest batch "%s" failed for an unknown reason.', $batch->id));
